libapp: add support for timers in the event loop code

+++ libs/libc/include/bits/timer_t.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif /* __cplusplus */
 
-typedef long timer_t;
+typedef int timer_t;
 
 #ifdef __cplusplus
 }
